% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_avar}
\alias{get_avar}
\title{Read an avar's remote representation by known box and known key.
Read the remote representation of an "asynchronous variable" ("avar").}
\usage{
get_avar(box = uuid(), key)
}
\arguments{
\item{box}{A string.}

\item{key}{A string.}
}
\value{
An R list
}
\description{
Read an avar's remote representation by known box and known key.
Read the remote representation of an "asynchronous variable" ("avar").
}

