% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_jobs}
\alias{get_jobs}
\title{Read a list of keys for tasks that have a known box and known status.}
\usage{
get_jobs(box = uuid(), status = "waiting")
}
\arguments{
\item{box}{A string.}

\item{status}{A string.}
}
\value{
An R list
}
\description{
Read a list of keys for tasks that have a known box and known status.
}

