% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set_avar}
\alias{set_avar}
\title{Write an avar by known box and known key.}
\usage{
set_avar(box = uuid(), key = uuid(), status = NULL, val = NULL)
}
\arguments{
\item{box}{A string.}

\item{key}{A string.}

\item{status}{A string.}

\item{val}{Any R data type.}
}
\description{
Write an avar by known box and known key.
}

