% Generated by roxygen2 (4.0.2): do not edit by hand
\name{submit}
\alias{submit}
\title{Submit a job to be distributed to a remote execution context.}
\usage{
submit(x = NULL, f, box = uuid(), env = environment(f))
}
\arguments{
\item{x}{Input data}

\item{f}{A function.}

\item{box}{A string.}

\item{env}{An environment.}
}
\value{
The results of `f(x)`
}
\description{
Submit a job to be distributed to a remote execution context.
}

